requirements.txt
wheel
dramatiq==1.15.0
flake8==5.0.4
Flask==2.3.3
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
gunicorn==20.1.0
pika==1.3.1
pytest==7.2.0
python-dotenv==0.21.0
rabbitmq==0.2.0
redis==4.5.5
